Exam practice small tasks and tests

+++ arrays/src/KrisztiFourthTask.java
@@ -5,11 +5,11 @@ public class KrisztiFourthTask {
     }
 
     // (-1) * 1 + 1 * 2 + 0 * 3 + (-1) * 4 = -1 + 2 -4 = -3
-    public static int countStrangenumbers(int[] numbers) {
+    public static long countStrangenumbers(int[] numbers) {
         if (numbers.length == 0) {
             return 0;
         }
-        int sum = 0;
+        long sum = 0;
         for (int i = 0; i < numbers.length; i++) {
             if (i % 3 == 0) {
                 sum -= numbers[i];
